DEBUGGED the MOD crash
MOD was crashing with the case:
LET A 10
LET B 5
MOD A B 

Giving a big red dot on the microbit. I introduced a safe_mod routine to check for divide zero, reworked the assembly code and added some debug statements which are now commented out for later use. 

Works now!